[tutorial] introducing database code in test-driven manner
However, as I wrote this, I found I could no longer follow the tutorial as
written: it was simply too hard to maintain the original code and test it
and not have the complexity get out of hand.  After discussing this issue
with Marko Jozik at the 2014 Austrian Perl Workshop I decided to rewrite
this section and change how the application is introduced at the very
beginning.  This commit is merely to save the current state of the document
should it be necessary to use this material at some time in the future.